♻️ LabelGroupコンポーネントをBaseSidebarGroupを元に作成するよう修正。
